% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_policy_rf}
\alias{get_policy_rf}
\title{Get RF-predicted policy scores (2014--2021)}
\usage{
get_policy_rf(year = NULL, cols = NULL)
}
\arguments{
\item{year}{Integer vector of survey years to keep.}

\item{cols}{Character vector of column names to return.}
}
\value{
A data frame containing RF-predicted policy scores.
}
\description{
Get RF-predicted policy scores (2014--2021)
}
