% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy_rf}
\alias{policy_rf}
\title{Random-forest predicted policy scores by year (2014--2021)}
\format{
A data frame with \code{case_id}, \code{year}, and policy columns
suffixed with \code{_rf}.
}
\usage{
policy_rf
}
\description{
Predicted policy preference scores from the RF model.
}
\keyword{datasets}
