% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_policy_lm}
\alias{get_policy_lm}
\title{Get LM-predicted policy scores (2014--2021)}
\usage{
get_policy_lm(year = NULL, cols = NULL)
}
\arguments{
\item{year}{Integer vector of survey years to keep.}

\item{cols}{Character vector of column names to return.}
}
\value{
A data frame containing LM-predicted policy scores.
}
\description{
Get LM-predicted policy scores (2014--2021)
}
