% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy_svr}
\alias{policy_svr}
\title{SVR predicted policy scores by year (2014--2021)}
\format{
A data frame with \code{case_id}, \code{year}, and policy columns
suffixed with \code{_svr}.
}
\usage{
policy_svr
}
\description{
Predicted policy preference scores from the SVR model.
}
\keyword{datasets}
